+++ matrix.cpp
@@ -1,14 +1,14 @@
 #include "matrix.h"
 
-const Duo<int> &Matrix::GetSize() const { return size; }
+const Duo<size_t> &Matrix::GetSize() const { return size; }
 
-Matrix::Matrix(int mxsizex, int mxsizey, int **mxgrid) : size(Duo<int>(mxsizex, mxsizey)), grid(nullptr)
+Matrix::Matrix(int mxsizex, int mxsizey, int **mxgrid) : size(Duo<size_t>(mxsizex, mxsizey)), grid(nullptr)
 {
     grid = new int *[size.x];
-    for (int s = 0; s < size.x; s++)
+    for (size_t s = 0; s < size.x; s++)
     {
         grid[s] = new int[size.y];
-        for (int o = 0; o < size.y; o++)
+        for (size_t o = 0; o < size.y; o++)
         {
             grid[s][o] = mxgrid[s][o];
         }
@@ -20,10 +20,10 @@ Matrix::Matrix(const Matrix &mx) : size(mx.GetSize())
     if (size.x != 0)
     {
         grid = new int *[size.x];
-        for (int s = 0; s < size.x; s++)
+        for (size_t s = 0; s < size.x; s++)
         {
             grid[s] = new int[size.y];
-            for (int o = 0; o < size.y; o++)
+            for (size_t o = 0; o < size.y; o++)
             {
                 grid[s][o] = mx.grid[s][o];
             }
@@ -33,16 +33,16 @@ Matrix::Matrix(const Matrix &mx) : size(mx.GetSize())
 
 void Matrix::operator=(const Matrix &mx)
 {
-    for (int i = 0; i < size.x; i++)
+    for (size_t i = 0; i < size.x; i++)
         delete[] grid[i];
     delete[] grid;
 
     size = mx.GetSize();
     grid = new int *[size.x];
-    for (int s = 0; s < size.x; s++)
+    for (size_t s = 0; s < size.x; s++)
     {
         grid[s] = new int[size.y];
-        for (int o = 0; o < size.y; o++)
+        for (size_t o = 0; o < size.y; o++)
         {
             grid[s][o] = mx.grid[s][o];
         }
@@ -51,7 +51,7 @@ void Matrix::operator=(const Matrix &mx)
 
 Matrix::~Matrix()
 {
-    for (int i = 0; i < size.x; i++)
+    for (size_t i = 0; i < size.x; i++)
         delete[] grid[i];
 
     delete[] grid;
